Extract some stuff from build into mypy_hacks
This works around our lack of support for accepting *args.
